Use AppLifecyleAnalytics directly
